requirements.txt
flask
pycryptodome
python-dotenv
awslambdaric